requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.7.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.12.4
dnspython==1.16.0
mongoengine==0.23.1
pymongo==3.12.0
pytz==2021.1